requirements.txt.txt
transformers
datasets
torch
scikit-learn
pandas
fastapi
uvicorn
streamlit
apscheduler
newsapi-python
requests
python-dotenv
